Candy implementation
I implemented the candies using an animation to change the 'Darkness' sprite and make the lantern intensity look bigger. I did that very quickly so we have to add more intensity levels and redo the sprites. They have to be more centered and of better resolution.

Note : You have to click "E" to pick up the candy. We can change that later